requirements.txt
Flask==2.0.1
pymongo[srv]==3.12.0
redis==3.5.3
Werkzeug==2.0.1
google-generativeai==0.3.0
python-dotenv==1.0.0
gunicorn==20.1.0
google-cloud-secret-manager==2.16.1